Undo mindless commenting of writing in unit tests, likely done by unsupervised agents. Undo or fix a few other things.
Entire-Checkpoint: 01M2427GN9106RWXGEEAWE35TF
