+++ gyrokinetic/unit/ctest_bc_twistshift.c
@@ -306,8 +306,8 @@ test_bc_twistshift_3x_fig6_wcells(const int *cells, enum gkyl_edge_loc edge,
       .basis_type = basis.id
     }
   );
-  // if (write_f)
-  //   gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x_fig6_do.gkyl");
+  if (write_f)
+    gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x_fig6_do.gkyl");
 
   // Create a range only extended in bc_dir.
   struct gkyl_range update_rng;
@@ -563,8 +563,8 @@ test_bc_twistshift_3x2v_fig6_wcells(const int *cells, enum gkyl_edge_loc edge,
       .basis_type = basis.id
     }
   );
-  // if (write_f)
-  //   gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x2v_fig6_do.gkyl");
+  if (write_f)
+    gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x2v_fig6_do.gkyl");
 
   // Create a range only extended in bc_dir.
   struct gkyl_range update_rng;
@@ -963,8 +963,8 @@ test_bc_twistshift_3x_fig11_wcells(const int *cells, enum gkyl_edge_loc edge,
       .basis_type = basis.id
     }
   );
-  // if (write_f)
-  //   gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x_fig11_do.gkyl");
+  if (write_f)
+    gkyl_grid_sub_array_write(&grid, &local, mt, distf_ho, "ctest_bc_twistshift_3x_fig11_do.gkyl");
 
   // Create a range only extended in bc_dir.
   struct gkyl_range update_rng;
